Moving EVERY file to a git repository
All of the files included came from an old directory on my computer when I was using these for handling iLearn student submissions for my CSC job at TTU in the Spring of 2024.
There is a ton of work that needs to be done, and I want to move this to a GitHub repo to make it a little easier to keep track of what is going on.